requirements.txt
timm
tqdm
numpy
openai
mutagen
torch
torchaudio
torchvision
pandas
pillow
openpyxl
moviepy==1.0.3
matplotlib
pyinstaller
google-genai
opencv-python
python-dotenv
